Fix the "Clear Color" button
Make it so the Clear Color button actually works. Previously the code
would set it to Yellow Orb, without even clearing where Yellow Orb was
at the time.

It now wipes all POI from the current room and recolors it to a
"standard room".

Also re-enables the "Clear Color" button as it was being disabled,
likely because it didn't work :wink:.